+++ src/binding/unbindModule.ts
@@ -1,5 +1,5 @@
 import { NewableModule } from "../types/Module";
-import { IS_MODULE_KEY, MODULE_IS_BINDED_KEY } from "../constants";
+import { IS_MODULE_KEY, MODULE_IS_BOUND_KEY } from "../constants";
 import messagesMap from "../messages/messagesMap";
 import isNewable from "../validation/isNewable";
 import { getModuleMetadata } from "../metadata/getModuleMetadata";
@@ -11,9 +11,9 @@ export default async function unbindModule(Module: NewableModule) {
     (item) => isNewable(item)
   ) as NewableModule[];
 
-  if (metadata.isBinded) {
+  if (metadata.isBound) {
     await metadata.container.unbindAllAsync();
-    defineMetadata(MODULE_IS_BINDED_KEY, false, Module.prototype);
+    defineMetadata(MODULE_IS_BOUND_KEY, false, Module.prototype);
   }
 
   for (const item of newableModulesImported) {
@@ -22,7 +22,7 @@ export default async function unbindModule(Module: NewableModule) {
     if (isModule) {
       await unbindModule(item);
     } else {
-      console.warn(messagesMap.notAModuleUnbinded(item.name));
+      console.warn(messagesMap.notAModuleUnbound(item.name));
     }
   }
 }