+++ src/constants.ts
@@ -3,7 +3,7 @@ import ModuleMetadata from "./types/ModuleMetadata";
 
 export const IS_MODULE_KEY: keyof ModuleMetadata = "isModule";
 
-export const MODULE_IS_BINDED_KEY: keyof ModuleMetadata = "isBinded";
+export const MODULE_IS_BOUND_KEY: keyof ModuleMetadata = "isBound";
 
 export const MODULE_METADATA_KEYS = keys<ModuleMetadata>();