+++ src/importing/importModule.ts
@@ -2,7 +2,7 @@ import { Newable } from "../types";
 import InversifySugar from "../inversifySugar/InversifySugar";
 import messagesMap from "../messages/messagesMap";
 import bindImportsToModule from "./bindImportsToModule";
-import { MODULE_IS_BINDED_KEY } from "../constants";
+import { MODULE_IS_BOUND_KEY } from "../constants";
 import bindProviderToContainer from "../binding/bindProviderToContainer";
 import { NewableModule } from "../types/Module";
 import { getModuleMetadata } from "../metadata/getModuleMetadata";
@@ -38,7 +38,7 @@ export default function importModule(
 function importRootModule(Module: NewableModule) {
   const metadata = getModuleMetadata(Module);
 
-  if (!metadata.isBinded) {
+  if (!metadata.isBound) {
     bindImportsToModule(Module, metadata.imports);
 
     for (const provider of metadata.providers.concat(
@@ -47,9 +47,9 @@ function importRootModule(Module: NewableModule) {
       bindProviderToContainer(provider, InversifySugar.globalContainer);
     }
 
-    InversifySugar.onModuleBinded(metadata.container, metadata, Module);
+    InversifySugar.onModuleBound(metadata.container, metadata, Module);
 
-    defineMetadata(MODULE_IS_BINDED_KEY, true, Module.prototype);
+    defineMetadata(MODULE_IS_BOUND_KEY, true, Module.prototype);
   }
 }
 
@@ -57,7 +57,7 @@ function importChildModule(Module: NewableModule): ExportedProviderRef[] {
   const metadata = getModuleMetadata(Module);
   const exportedProviderRefs: ExportedProviderRef[] = [];
 
-  if (!metadata.isBinded) {
+  if (!metadata.isBound) {
     bindImportsToModule(Module, metadata.imports);
 
     for (const provider of metadata.globalProviders) {
@@ -68,9 +68,9 @@ function importChildModule(Module: NewableModule): ExportedProviderRef[] {
       metadata.container.bindProvider(provider);
     }
 
-    InversifySugar.onModuleBinded(metadata.container, metadata, Module);
+    InversifySugar.onModuleBound(metadata.container, metadata, Module);
 
-    defineMetadata(MODULE_IS_BINDED_KEY, true, Module.prototype);
+    defineMetadata(MODULE_IS_BOUND_KEY, true, Module.prototype);
   }
 
   for (const exportedProvider of metadata.exports) {