+++ src/inversifySugar/inversifySugarOptions.ts
@@ -3,7 +3,7 @@ import { InversifySugarOptions } from "../types";
 export const defaultInversifySugarOptions: InversifySugarOptions = {
   debug: false,
   defaultScope: "Transient",
-  onModuleBinded: undefined,
+  onModuleBound: undefined,
 };
 
 const inversifySugarOptions: InversifySugarOptions = Object.assign(