+++ src/validation/isValidProvide.ts
@@ -1,5 +1,9 @@
 const isValidProvide = (data: unknown): boolean => {
-  return typeof data === "string" || typeof data === "symbol";
+  return (
+    typeof data === "string" ||
+    typeof data === "symbol" ||
+    (typeof data === "function" && /^\s*class\s+/.test(data.toString()))
+  );
 };
 
 export default isValidProvide;